+++ CHANGELOG.md
@@ -2,3 +2,5 @@ Change Log
 ==========
 
 For changes to the Kotlin library, see [kotlin/CHANGELOG.md](kotlin/CHANGELOG.md).
+
+For changes to the Swift library, see [swift/CHANGELOG.md](swift/CHANGELOG.md).