+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Change Log
 ==========
 
+## Version 0.14.1
+
+_2019-4-18_
+
+### Kotlin
+
+ * Fix for WorkflowHost.Factory not using baseContext. ([#306](https://github.com/square/workflow/pull/306))
+
+### Swift
+
+ * Kotlin-only release, no code changes.
+
 ## Version 0.14.0
 
 _2019-4-16_